Optimizations + code clean up
