--- src/components/ConfigMapSecret/ConfigMap/ConfigMap.scss
@@ -1,135 +1,0 @@
-.white-card__header {
-    margin-bottom: 24px;
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-}
-
-.info__container--configmap {
-    padding: 12px;
-}
-
-.white-card__config-map {
-    .form-row {
-        margin-bottom: 24px;
-    }
-
-    .form__label {
-        font-size: 12px;
-        color: var(--N700);
-    }
-
-    .form-row__tab {
-        width: 100%;
-        display: grid;
-        grid-template-columns: 1fr 1fr;
-        border: 1px solid var(--N200);
-        border-radius: 4px;
-        overflow: hidden;
-
-        .form__label {
-            grid-column: 1 / span 2;
-        }
-
-        .form__tab {
-            padding: 12px 16px;
-            height: 48px;
-            cursor: pointer;
-            border: unset;
-            border-radius: 0;
-
-            .tab_title {
-                color: var(--N900);
-                font-size: 13px;
-            }
-
-            .tab__selector {
-                height: 16px;
-                width: 16px;
-                border: 1px solid var(--N200);
-                border-radius: 50%;
-                margin-right: 4px;
-            }
-
-            &:not(:first-child) {
-                border-left: 1px solid var(--N200);
-                border-radius: 0;
-            }
-
-            &.active {
-                color: var(--B500);
-                font-weight: 600;
-
-                .tab__selector {
-                    background: radial-gradient(var(--B500) 0%, var(--B500) 40%, white 41%, white 100%);
-                }
-            }
-        }
-
-        &.disabled {
-            .form__tab:not(.active) {
-                background: var(--N100);
-            }
-        }
-    }
-
-    .add-parameter {
-        img {
-            margin-right: 4px;
-            height: 20px;
-            width: 20px;
-        }
-
-        .edit {
-            margin-right: 16px;
-            color: --N500;
-        }
-
-        color: var(--B500);
-        text-align: left;
-        margin-left: auto;
-    }
-
-    .uncollapse__delete {
-        width: 72px;
-        justify-content: flex-end;
-
-        svg {
-            margin-right: auto;
-        }
-    }
-
-    .unknown {
-        height: 20px;
-        width: 20px;
-        margin-left: 4px;
-        cursor: auto;
-        background-size: contain;
-    }
-}
-
-.configuration-list__info {
-    display: flex;
-    align-items: center;
-}
-
-.form-row__select-external-type {
-    justify-content: flex-start;
-
-    .edit {
-        margin-left: auto;
-    }
-
-    .popup-button {
-        width: 50%;
-    }
-
-    .select-button {
-        height: 40px;
-        padding-left: 10px;
-    }
-}
-
-.form-label__select-external-type {
-    margin: 20px 0 10px 0;
-}
