--- src/components/ConfigMapSecret/ConfigMapSecret.utils.ts
@@ -1,10 +1,0 @@
-export const getLabel = (isOverrideView, defaultData, data): string => {
-    if (isOverrideView) {
-        if (defaultData) {
-            return data ? 'Overridden' : 'Inheriting'
-        } else {
-            return 'env'
-        }
-    }
-    return ''
-}
