+++ src/components/ConfigMapSecret/Constants.ts
@@ -27,3 +27,10 @@ export enum CM_SECRET_STATE {
     OVERRIDDEN = 'Overridden',
     ENV = 'Env',
 }
+
+export const SECRET_TOAST_INFO = {
+  BOTH_STORE_AVAILABLE: 'Please use either secretStore or secretStoreRef',
+  CHECK_KEY_SECRET_KEY: 'Please check key and secretKey',
+  BOTH_STORE_UNAVAILABLE: 'Please provide secretStore or secretStoreRef',
+  CHECK_KEY_NAME: 'Please check key and name'
+}
