+++ src/components/ConfigMapSecret/Secret/SecretList.tsx
@@ -4,7 +4,6 @@ import { useParams } from 'react-router'
 import { getAppChartRefForAppAndEnv } from '../../../services/service'
 import { DOCUMENTATION } from '../../../config'
 import '../ConfigMap.scss'
-import { decode } from '../../../util/Util'
 import { ConfigMapSecretContainer } from '../ConfigMapSecret.components'
 import InfoIconWithTippy from '../InfoIconWithTippy'
 import { ConfigMapListProps } from '../Types'
