--- src/components/ConfigMapSecret/Secret/constants.ts
@@ -1,6 +1,0 @@
-export const SECRET_TOAST_INFO = {
-    BOTH_STORE_AVAILABLE: 'Please use either secretStore or secretStoreRef',
-    CHECK_KEY_SECRET_KEY: 'Please check key and secretKey',
-    BOTH_STORE_UNAVAILABLE: 'Please provide secretStore or secretStoreRef',
-    CHECK_KEY_NAME: 'Please check key and name'
-}
