+++ src/components/ConfigMapSecret/service.ts
@@ -18,12 +18,12 @@ export function deleteEnvConfigMap(id, appId, envId, name) {
     return trash(`${Routes.APP_CREATE_ENV_CONFIG_MAP}/${appId}/${envId}/${id}?name=${name}`)
 }
 
-export function overRideConfigMap(id, appId, environmentId, configData){
-    return post(`${ Routes.APP_CREATE_ENV_CONFIG_MAP }`, {
+export function overRideConfigMap(id, appId, environmentId, configData) {
+    return post(`${Routes.APP_CREATE_ENV_CONFIG_MAP}`, {
         id,
         appId,
         environmentId,
-        configData
+        configData,
     })
 }
 
@@ -35,15 +35,27 @@ export function getConfigMapList(appId, envId?) {
     }
 }
 
-export function deleteSecret(id, appId, name){
-  return trash(`${Routes.APP_CREATE_SECRET}/${appId}/${id}?name=${name}`)
+export function updateSecret(id, appId, configData) {
+    return post(`${Routes.APP_CREATE_SECRET}`, {
+        ...(id && { id }),
+        appId,
+        configData: [configData],
+    })
+}
+
+export function getSecretKeys(id, appId, name) {
+    return get(`${Routes.APP_CREATE_SECRET}/edit/${appId}/${id}?name=${name}`)
+}
+
+export function deleteSecret(id, appId, name) {
+    return trash(`${Routes.APP_CREATE_SECRET}/${appId}/${id}?name=${name}`)
 }
 
 export function deleteEnvSecret(id, appId, envId, name) {
-  return trash(`${Routes.APP_CREATE_ENV_SECRET}/${appId}/${envId}/${id}?name=${name}`)
+    return trash(`${Routes.APP_CREATE_ENV_SECRET}/${appId}/${envId}/${id}?name=${name}`)
 }
 
-export function unlockEnvSecret(id, appId, envId, name){
+export function unlockEnvSecret(id, appId, envId, name) {
     return get(`${Routes.APP_CREATE_ENV_SECRET}/edit/${appId}/${envId}/${id}?name=${name}`)
 }
 
@@ -55,13 +67,11 @@ export function getSecretList(appId, envId?) {
     }
 }
 
-
 export function overRideSecret(id, appId, environmentId, configData) {
     return post(`${Routes.APP_CREATE_ENV_SECRET}`, {
         id,
         appId,
         environmentId,
-        configData
+        configData,
     })
 }
-
